+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/App.tsx
@@ -1,5 +1,5 @@
 import "./App.css";
-import AuthProvider from "./state-management/AuthProvider";
+import AuthProvider from "./state-management/auth/AuthProvider";
 import HomePage from "./state-management/HomePage";
 import NavBar from "./state-management/NavBar";
 import TaskProvider from "./state-management/tasks/TaskProvider";
