--- Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/AuthProvider.tsx
@@ -1,18 +1,0 @@
-import { ReactNode, useReducer } from "react";
-import authReducer from "./reducers/authReducer";
-import AuthContext from "./contexts/authContext";
-
-interface Props {
-  children: ReactNode;
-}
-const AuthProvider = ({ children }: Props) => {
-  const [user, dispatch] = useReducer(authReducer, "");
-
-  return (
-    <AuthContext.Provider value={{ user, dispatch }}>
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export default AuthProvider;
