+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/NavBar.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import LoginStatus from "./LoginStatus";
+import LoginStatus from "./auth/LoginStatus";
 import TaskContext from "./tasks/taskContext";
 
 const NavBar = () => {
