+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/auth/authContext.ts
@@ -1,5 +1,5 @@
 import React, { Dispatch } from "react";
-import { AuthAction } from "../reducers/authReducer";
+import { AuthAction } from "./AuthProvider";
 
 interface AuthContextType {
   user: string;
