+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/counter/Counter.tsx
@@ -1,5 +1,5 @@
-import { useReducer, useState } from "react";
-import counterReducer from "./reducers/counterReducer";
+import { useReducer } from "react";
+import counterReducer from "./counterReducer";
 
 const Counter = () => {
   // const [value, setValue] = useState(0);
