+++ Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/counter/counterReducer.ts
