--- Mahmoud/Global-State-Management/react-course-part2-starter-main/src/state-management/hooks/useAuth.ts
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import AuthContext from "../contexts/authContext";
-
-const useAuth = () => useContext(AuthContext);
-
-export default useAuth;
